feat: weather forecast - move feat to internal
